Update build chain to create proper .app releases for MacOS with SDL bundled correctly

+++ gameSource/LivingLifePage.cpp
@@ -51,6 +51,7 @@
 #include <stdlib.h>//#include <math.h>
 #include <string>
 #include <fstream>
+#include <sstream>
 
 
 #define OHOL_NON_EDITOR 1
@@ -8147,9 +8148,9 @@ void LivingLifePage::draw( doublePair inViewCenter,
     setDrawColor( 1, 1, 1, 1 );
 
     int gridCenterX = 
-        lrintf( lastScreenViewCenter.x / CELL_D ) - mMapOffsetX + mMapD/2;
+        (int)lrintf( lastScreenViewCenter.x / CELL_D ) - mMapOffsetX + mMapD/2;
     int gridCenterY = 
-        lrintf( lastScreenViewCenter.y / CELL_D ) - mMapOffsetY + mMapD/2;
+        (int)lrintf( lastScreenViewCenter.y / CELL_D ) - mMapOffsetY + mMapD/2;
     
     //FOV
     // SIDE NOTE:  These 4 variables control how far items should be rendered, separately from biome drawing
@@ -9002,8 +9003,8 @@ void LivingLifePage::draw( doublePair inViewCenter,
 
     if( vogModeActuallyOn ) {
         GridPos viewingGridPos = {
-            lrintf( lastScreenViewCenter.x / CELL_D ),
-            lrintf( lastScreenViewCenter.y / CELL_D )
+            (int)lrintf( lastScreenViewCenter.x / CELL_D ),
+            (int)lrintf( lastScreenViewCenter.y / CELL_D )
             };
         FloatColor tileColor = {0.5, 0.0, 1.0, 1.0};
         if( vogScrollingMode ) tileColor = {1.0, 0.0, 1.0, 1.0};
@@ -25736,8 +25737,8 @@ void LivingLifePage::pointerMove( float inX, float inY ) {
             }
 
         GridPos viewingGridPos = {
-            lrintf( lastScreenViewCenter.x / CELL_D ),
-            lrintf( lastScreenViewCenter.y / CELL_D )
+            (int)lrintf( lastScreenViewCenter.x / CELL_D ),
+            (int)lrintf( lastScreenViewCenter.y / CELL_D )
             };
 
         if( // maxChunkDimension/2 is the border of loaded area
@@ -28329,8 +28330,8 @@ void LivingLifePage::keyDown( unsigned char inASCII ) {
                 if( vogMode ) {
 
                     GridPos viewingGridPos = {
-                        lrintf( lastScreenViewCenter.x / CELL_D ),
-                        lrintf( lastScreenViewCenter.y / CELL_D )
+                        (int)lrintf( lastScreenViewCenter.x / CELL_D ),
+                        (int)lrintf( lastScreenViewCenter.y / CELL_D )
                         };
 
                     if( viewingGridPos.x != lrint( vogPos.x ) || 
@@ -28958,8 +28959,8 @@ void LivingLifePage::keyDown( unsigned char inASCII ) {
                             // at the wrong position
                             
                             GridPos viewingGridPos = {
-                                lrintf( lastScreenViewCenter.x / CELL_D ),
-                                lrintf( lastScreenViewCenter.y / CELL_D )
+                                (int)lrintf( lastScreenViewCenter.x / CELL_D ),
+                                (int)lrintf( lastScreenViewCenter.y / CELL_D )
                                 };
 
                             if( viewingGridPos.x != lrint( vogPos.x ) || 
@@ -29071,8 +29072,8 @@ void LivingLifePage::specialKeyDown( int inKeyCode ) {
         if( posOffset.x != 0 || posOffset.y != 0 ) {
             
             GridPos newPos;
-            newPos.x = lrintf( lastScreenViewCenter.x / CELL_D );
-            newPos.y = lrintf( lastScreenViewCenter.y / CELL_D );
+            newPos.x = (int)lrintf( lastScreenViewCenter.x / CELL_D );
+            newPos.y = (int)lrintf( lastScreenViewCenter.y / CELL_D );
             
             newPos.x += posOffset.x;
             newPos.y += posOffset.y;
